CRS-652 (#124)
* CRS-652
Added the example for CRS-652 which was throwing an exception.
Created a workaround for the problem until it can be resolved by using per sentence adjustments. Put in some logging, so we can keep track. Added comments and todo so we know where to unpick the problem in the future.
Suppressed the unchecked cast, to avoid warning.

* Fixed small linting issue