CRS-1958: Handle gap between offence committed date and sentence date for list D offences
If the offence was committed before PCSC date but sentenced after and the offence appears on list D and is 7 years or more then it should. These offence codes were the ones applicable at the time the offence was committed so the code itself does not appear on list D although the offence does.