CRS-1942 SDS early release support (#674)
* CRS-1942: Set early release if not excluded

Set early release track if not excluded and the feature toggle is on. Turned on feature toggle in alt envs and set default back to 50 with relevant overrides for alt params.

* CRS-1942: Use appropriate release arrangement for consec chain

The consecutive release was assuming only 2 possible release arrangements. Now supports multiple release arrangements. They are applied in the order the release arrangement appears in the chain.