crs-2273-hdc12-consecutive-sentences (#1096)
* crs-2273-hdc12-consecutive-sentences

For consecutive sentence with sentences in the chain imposed before and after the HDC12 rollout date, default the sentence start date to HDC date where the imposed sentence date is prior to HDC12.

* crs-2283-enable-concurrent-consecutive

For consecutive sentence with sentences in the chain imposed before and after the HDC12 rollout date, default the sentence start date to HDC date where the imposed sentence date is prior to HDC12.

* Add additional tests cases

* crs-2273-hdc12-consecutive-sentences

Move pre HDC defaulting logic to applyHdc365Rules method.