CRS-2386: Update HDC rules regarding UAL
Reworked HdcedCalculator so we keep track of an unadjusted date as well as the proper HDCED. Unadjusted in this context means only the deductions have been applied. These unadjusted dates are then used to work out which HDC rules apply (180, 365 or commencement date).
When there is UAL an additional unadjusted date is checked to see if with UAL applied it is still before the commencement date. Previously defaulting to the commencement date and then applying UAL could mean more than the required UAL is applied.