Initial support for ERS30 (#1159)
* CRS-2381: Initial support for ERS30

When the feature toggle is enabled calculate both ERS30 and ERS50 dates. If ERS50 excluding ADAs + RADAs is before commencement, then retain ERS50. If ERS30 excluding ADAs + RADAs is before commencement, default to the commencement date. Otherwise, use the ERS30.

* CRS-2381: Remove test for acceptance criteria L

Pending further investigation into whether ERSED should be calculated in this case or not as it is generated both before and after these changes.