Crs 2480 pss repeal (#1211)
* crs-2480-pss-repeal

Default ANY Calculated OR Retrieved TUSEDs that are ON OR AFTER the date of repeal to the day before.

Retain ANY Calculated or Retrieved TUSEDs that are BEFORE the date of repeal.

Not Calculate any NEW TUSED’s for sentences imposed ON or AFTER the date of repeal.

* crs-2480-pss-repeal

Default ANY Calculated OR Retrieved TUSEDs that are ON OR AFTER the date of repeal to the day before.

Retain ANY Calculated or Retrieved TUSEDs that are BEFORE the date of repeal.

Not Calculate any NEW TUSED’s for sentences imposed ON or AFTER the date of repeal.

* crs-2480-pss-repeal

remove test method

* crs-2480-pss-repeal

remove test method

* crs-2480-pss-repeal

Refactor naming convention resulting from code review.