CRS-2536: Fix incorrect hint on CRDS calcs (#1262)
We were showing the "Manually overridden" hint on CRDS bookings which can only be overridden by genuine override. This is because we were looking at the sentenceDetail endpoint of Prison API using the booking id which is for the latest calculation only.

Changed to not bother loading sentence detail for any CRDS calc. A further change is required to fix this for historical NOMIS calcs but it requires a change to Prison API.