crs-2465-revocation-date-check (#1307)
Exclude FTR56 sentences where the revocation date is on the date of FTR56 commencement. The system cannot be sure if the sentence is new or converted from older FTR type in Nomis.

Sentence is eligible for Tranche if return to custody date falls on the same date as Tranche commencement.