crs-2556-ual-rules-for-ftr56 (#1309)
* crs-2556-ual-rules-for-ftr56

Where the escape/abscond is on or before day 56 of the FTR56 custodial period, the UAL is applied to both the PRRD and the SLED. Where the escape/abscond is after day 56 of the FTR56 custodial period, the UAL is only applied to the SLED. The determination of whether the converted FTR56 PRRD falls before tranche commencement date should be performed after any UAL has been applied to the PRRD i.e. the UAL adjusted PRRD should be considered for FTR56 defaulting rules.

Track post return to custody date UAL within the adjustments data class for FTR56.
Reference the days when returning the PRRD and default where necessary.

* Update SentenceCalculation.kt (#1310)

---------

Co-authored-by: ldlharper <luke.harper@digital.justice.gov.uk>